Move selection details code into UI module
